@@ -1,15 +1,19 @@
 #!/usr/bin/env python3
 
-import setuptools
 from pathlib import Path
-from typing import cast, Dict
+from typing import cast
+
+import setuptools
 
 package_name = 'vskernels'
 
-exec(Path(f'{package_name}/_metadata.py').read_text(), meta := cast(Dict[str, str], {}))
+exec(Path(f'{package_name}/_metadata.py').read_text(), meta := cast(dict[str, str], {}))
 
 readme = Path('README.md').read_text()
 requirements = Path('requirements.txt').read_text()
+
+# stubs comand
+# vsgenstubs4 std resize descale fmtc placebo -o stubs
 
 
 setuptools.setup(
@@ -28,7 +32,7 @@ setuptools.setup(
         'Contact': 'https://discord.gg/qxTxVJGtst',
     },
     install_requires=requirements,
-    python_requires='>=3.8',
+    python_requires='>=3.10',
     packages=[
         package_name, f'{package_name}.kernels'
     ],
@@ -41,7 +45,7 @@ setuptools.setup(
         "Intended Audience :: Developers",
         "Intended Audience :: Other Audience",
 
-        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.10",
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
         "Typing :: Typed",
