@@ -1,11 +1,15 @@
 from __future__ import annotations
 
-from typing import Any, List, Tuple, overload
+from typing import Any, overload
 
 import vapoursynth as vs
+from vstools import HoldsVideoFormatT, MatrixT, VideoFormatT, get_format, inject_self
 
-from ..types import MatrixT, VideoFormatT
 from .abstract import Kernel
+
+__all__ = [
+    'Example'
+]
 
 core = vs.core
 
@@ -13,12 +17,15 @@ core = vs.core
 class Example(Kernel):
     """Example Kernel class for documentation purposes."""
 
-    def __init__(self, b: float = 0, c: float = 1/2, **kwargs: Any) -> None:
+    def __init__(self, b: float = 0, c: float = 1 / 2, **kwargs: Any) -> None:
         self.b = b
         self.c = c
         super().__init__(**kwargs)
 
-    def scale(self, clip: vs.VideoNode, width: int, height: int, shift: Tuple[float, float] = (0, 0)) -> vs.VideoNode:
+    @inject_self.cached
+    def scale(  # type: ignore[override]
+        self, clip: vs.VideoNode, width: int, height: int, shift: tuple[float, float] = (0, 0), **kwargs: Any
+    ) -> vs.VideoNode:
         """
         Perform a regular scaling operation.
 
@@ -32,10 +39,13 @@ class Example(Kernel):
         """
         return core.resize.Bicubic(
             clip, width, height, src_top=shift[0], src_left=shift[1],
-            filter_param_a=self.b, filter_param_b=self.c, **self.kwargs
+            filter_param_a=self.b, filter_param_b=self.c, **self.kwargs, **kwargs
         )
 
-    def descale(self, clip: vs.VideoNode, width: int, height: int, shift: Tuple[float, float] = (0, 0)) -> vs.VideoNode:
+    @inject_self.cached
+    def descale(  # type: ignore[override]
+        self, clip: vs.VideoNode, width: int, height: int, shift: tuple[float, float] = (0, 0), **kwargs: Any
+    ) -> vs.VideoNode:
         """
         Perform a regular descaling operation.
 
@@ -48,11 +58,13 @@ class Example(Kernel):
         :rtype:             ``VideoNode``
         """
         return core.descale.Debicubic(
-            clip, width, height, b=self.b, c=self.c, src_top=shift[0], src_left=shift[1]
+            clip, width, height, b=self.b, c=self.c, src_top=shift[0], src_left=shift[1], **kwargs
         )
 
-    def resample(
-        self, clip: vs.VideoNode, format: VideoFormatT, matrix: MatrixT | None = None, matrix_in: MatrixT | None = None
+    @inject_self.cached
+    def resample(  # type: ignore[override]
+        self, clip: vs.VideoNode, format: int | VideoFormatT | HoldsVideoFormatT,
+        matrix: MatrixT | None = None, matrix_in: MatrixT | None = None, **kwargs: Any
     ) -> vs.VideoNode:
         """
         Perform a regular resampling operation.
@@ -65,18 +77,18 @@ class Example(Kernel):
         :rtype:             ``VideoNode``
         """
         return core.resize.Bicubic(
-            clip, format=int(format),
+            clip, format=get_format(format).id,
             filter_param_a=self.b, filter_param_b=self.c,
-            matrix=matrix, matrix_in=matrix_in, **self.kwargs
+            matrix=matrix, matrix_in=matrix_in, **self.kwargs, **kwargs
         )
 
     @overload  # type: ignore
-    def shift(self, clip: vs.VideoNode, shift: Tuple[float, float] = (0, 0)) -> vs.VideoNode:
+    def shift(self, clip: vs.VideoNode, shift: tuple[float, float] = (0, 0), **kwargs: Any) -> vs.VideoNode:
         ...
 
     def shift(  # type: ignore
         self, clip: vs.VideoNode,
-        shift_top: float | List[float] = 0.0, shift_left: float | List[float] = 0.0
+        shift_top: float | list[float] = 0.0, shift_left: float | list[float] = 0.0, **kwargs: Any
     ) -> vs.VideoNode:
         """
         Perform a regular shifting operation.
@@ -91,5 +103,5 @@ class Example(Kernel):
         return core.resize.Bicubic(
             clip, src_top=shift_top, src_left=shift_left,  # type: ignore
             filter_param_a=self.b, filter_param_b=self.c,
-            **self.kwargs
+            **self.kwargs, **kwargs
         )
