@@ -1,3 +1,1488 @@
+# Stop pep8 from complaining (hopefully)
+# NOQA
+
+# Ignore Flake Warnings
 # flake8: noqa
 
-from .vapoursynth import *
+# Ignore coverage
+# (No coverage)
+
+# From https://gist.github.com/pylover/7870c235867cf22817ac5b096defb768
+# noinspection PyPep8
+# noinspection PyPep8Naming
+# noinspection PyTypeChecker
+# noinspection PyAbstractClass
+# noinspection PyArgumentEqualDefault
+# noinspection PyArgumentList
+# noinspection PyAssignmentToLoopOrWithParameter
+# noinspection PyAttributeOutsideInit
+# noinspection PyAugmentAssignment
+# noinspection PyBroadException
+# noinspection PyByteLiteral
+# noinspection PyCallByClass
+# noinspection PyChainedComparsons
+# noinspection PyClassHasNoInit
+# noinspection PyClassicStyleClass
+# noinspection PyComparisonWithNone
+# noinspection PyCompatibility
+# noinspection PyDecorator
+# noinspection PyDefaultArgument
+# noinspection PyDictCreation
+# noinspection PyDictDuplicateKeys
+# noinspection PyDocstringTypes
+# noinspection PyExceptClausesOrder
+# noinspection PyExceptionInheritance
+# noinspection PyFromFutureImport
+# noinspection PyGlobalUndefined
+# noinspection PyIncorrectDocstring
+# noinspection PyInitNewSignature
+# noinspection PyInterpreter
+# noinspection PyListCreation
+# noinspection PyMandatoryEncoding
+# noinspection PyMethodFirstArgAssignment
+# noinspection PyMethodMayBeStatic
+# noinspection PyMethodOverriding
+# noinspection PyMethodParameters
+# noinspection PyMissingConstructor
+# noinspection PyMissingOrEmptyDocstring
+# noinspection PyNestedDecorators
+# noinspection PynonAsciiChar
+# noinspection PyNoneFunctionAssignment
+# noinspection PyOldStyleClasses
+# noinspection PyPackageRequirements
+# noinspection PyPropertyAccess
+# noinspection PyPropertyDefinition
+# noinspection PyProtectedMember
+# noinspection PyRaisingNewStyleClass
+# noinspection PyRedeclaration
+# noinspection PyRedundantParentheses
+# noinspection PySetFunctionToLiteral
+# noinspection PySimplifyBooleanCheck
+# noinspection PySingleQuotedDocstring
+# noinspection PyStatementEffect
+# noinspection PyStringException
+# noinspection PyStringFormat
+# noinspection PySuperArguments
+# noinspection PyTrailingSemicolon
+# noinspection PyTupleAssignmentBalance
+# noinspection PyTupleItemAssignment
+# noinspection PyUnboundLocalVariable
+# noinspection PyUnnecessaryBackslash
+# noinspection PyUnreachableCode
+# noinspection PyUnresolvedReferences
+# noinspection PyUnusedLocal
+# noinspection ReturnValueFromInit
+
+
+from abc import abstractmethod
+from ctypes import Structure, c_void_p
+from enum import IntEnum
+from fractions import Fraction
+from inspect import Parameter, Signature
+from logging import NOTSET as LogLevelUnset
+from logging import Handler, LogRecord
+from types import MappingProxyType, TracebackType
+from typing import (
+    TYPE_CHECKING, Any, BinaryIO, Callable, ContextManager, Dict, Generic, Iterator, Literal, MutableMapping,
+    NamedTuple, NoReturn, Optional, Protocol, Sequence, Type, TypedDict, TypeVar, Union, overload, runtime_checkable
+)
+
+__all__ = [
+    # Versioning
+    '__version__', '__api_version__',
+
+    # Cython capsule
+    '__pyx_capi__',
+
+    # Enums and constants
+    'MessageType',
+        'MESSAGE_TYPE_DEBUG', 'MESSAGE_TYPE_INFORMATION', 'MESSAGE_TYPE_WARNING',
+        'MESSAGE_TYPE_CRITICAL', 'MESSAGE_TYPE_FATAL',
+
+    'FilterMode',
+        'fmParallel', 'fmParallelRequests', 'fmUnordered', 'fmFrameState',
+
+    'CoreCreationFlags',
+        'ccfEnableGraphInspection', 'ccfDisableAutoLoading', 'ccfDisableLibraryUnloading',
+
+    'MediaType',
+        'VIDEO', 'AUDIO',
+
+    'ColorFamily',
+        'UNDEFINED', 'GRAY', 'RGB', 'YUV',
+
+    'ColorRange',
+        'RANGE_FULL', 'RANGE_LIMITED',
+
+    'SampleType',
+        'INTEGER', 'FLOAT',
+
+    'PresetFormat',
+        'GRAY',
+        'GRAY8', 'GRAY9', 'GRAY10', 'GRAY12', 'GRAY14', 'GRAY16', 'GRAY32', 'GRAYH', 'GRAYS',
+        'RGB',
+        'RGB24', 'RGB27', 'RGB30', 'RGB36', 'RGB42', 'RGB48', 'RGBH', 'RGBS',
+        'YUV',
+        'YUV410P8',
+        'YUV411P8',
+        'YUV420P8', 'YUV420P9', 'YUV420P10', 'YUV420P12', 'YUV420P14', 'YUV420P16',
+        'YUV422P8', 'YUV422P9', 'YUV422P10', 'YUV422P12', 'YUV422P14', 'YUV422P16',
+        'YUV440P8',
+        'YUV444P8', 'YUV444P9', 'YUV444P10', 'YUV444P12', 'YUV444P14', 'YUV444P16', 'YUV444PH', 'YUV444PS',
+        'NONE',
+
+    'AudioChannels',
+        'FRONT_LEFT', 'FRONT_RIGHT', 'FRONT_CENTER',
+        'BACK_LEFT', 'BACK_RIGHT', 'BACK_CENTER',
+        'SIDE_LEFT', 'SIDE_RIGHT',
+        'TOP_CENTER',
+
+        'TOP_FRONT_LEFT', 'TOP_FRONT_RIGHT', 'TOP_FRONT_CENTER',
+        'TOP_BACK_LEFT', 'TOP_BACK_RIGHT', 'TOP_BACK_CENTER',
+
+        'WIDE_LEFT', 'WIDE_RIGHT',
+
+        'SURROUND_DIRECT_LEFT', 'SURROUND_DIRECT_RIGHT',
+
+        'FRONT_LEFT_OF_CENTER', 'FRONT_RIGHT_OF_CENTER',
+
+        'STEREO_LEFT', 'STEREO_RIGHT',
+
+        'LOW_FREQUENCY', 'LOW_FREQUENCY2',
+
+    'ChromaLocation',
+        'CHROMA_TOP_LEFT', 'CHROMA_TOP',
+        'CHROMA_LEFT', 'CHROMA_CENTER',
+        'CHROMA_BOTTOM_LEFT', 'CHROMA_BOTTOM',
+
+    'FieldBased',
+        'FIELD_PROGRESSIVE', 'FIELD_TOP', 'FIELD_BOTTOM',
+
+    'MatrixCoefficients',
+        'MATRIX_RGB', 'MATRIX_BT709', 'MATRIX_UNSPECIFIED', 'MATRIX_FCC',
+        'MATRIX_BT470_BG', 'MATRIX_ST170_M', 'MATRIX_YCGCO', 'MATRIX_BT2020_NCL', 'MATRIX_BT2020_CL',
+        'MATRIX_CHROMATICITY_DERIVED_NCL', 'MATRIX_CHROMATICITY_DERIVED_CL', 'MATRIX_ICTCP',
+
+    'TransferCharacteristics',
+        'TRANSFER_BT709', 'TRANSFER_UNSPECIFIED', 'TRANSFER_BT470_M', 'TRANSFER_BT470_BG', 'TRANSFER_BT601',
+        'TRANSFER_ST240_M', 'TRANSFER_LINEAR', 'TRANSFER_LOG_100', 'TRANSFER_LOG_316', 'TRANSFER_IEC_61966_2_4',
+        'TRANSFER_IEC_61966_2_1', 'TRANSFER_BT2020_10', 'TRANSFER_BT2020_12', 'TRANSFER_ST2084', 'TRANSFER_ARIB_B67',
+
+    'ColorPrimaries', 'PRIMARIES_BT709', 'PRIMARIES_UNSPECIFIED',
+        'PRIMARIES_BT470_M', 'PRIMARIES_BT470_BG', 'PRIMARIES_ST170_M', 'PRIMARIES_ST240_M', 'PRIMARIES_FILM',
+        'PRIMARIES_BT2020', 'PRIMARIES_ST428', 'PRIMARIES_ST431_2', 'PRIMARIES_ST432_1', 'PRIMARIES_EBU3213_E',
+
+    # Environment SubSystem
+    'Environment', 'EnvironmentData',
+
+    'EnvironmentPolicy',
+        'StandaloneEnvironmentPolicy',
+        'VSScriptEnvironmentPolicy',
+
+    'EnvironmentPolicyAPI',
+    'register_policy', 'has_policy',
+
+    'get_current_environment',
+
+    'VideoOutputTuple',
+    'clear_output', 'clear_outputs', 'get_outputs', 'get_output',
+
+    '_construct_type', '_construct_parameter', 'construct_signature',
+
+    # Logging
+    'PythonVSScriptLoggingBridge', 'LogHandle', 'Error',
+
+    # Functions
+    'FuncData', 'Func', 'VSMapValueCallbackData', 'FramePtr',
+    'Plugin', 'Function',
+
+    # Formats
+    'VideoFormat',
+
+    # Frames
+    'RawFrame', 'VideoFrame', 'AudioFrame',
+    'FrameProps',
+
+    # Nodes
+    'RawNode', 'VideoNode', 'AudioNode',
+
+    'Core', '_CoreProxy', 'core',
+
+    # Inspection API [UNSTABLE API]
+    '_try_enable_introspection'
+]
+
+
+###
+# Typing
+
+T = TypeVar('T')
+S = TypeVar('S')
+
+SingleAndSequence = Union[T, Sequence[T]]
+
+
+@runtime_checkable
+class SupportsString(Protocol):
+    @abstractmethod
+    def __str__(self) -> str:
+        ...
+
+
+DataType = Union[str, bytes, bytearray, SupportsString]
+
+_VapourSynthMapValue = Union[
+    SingleAndSequence[int],
+    SingleAndSequence[float],
+    SingleAndSequence[DataType],
+    SingleAndSequence['VideoNode'],
+    SingleAndSequence['VideoFrame'],
+    SingleAndSequence['AudioNode'],
+    SingleAndSequence['AudioFrame'],
+    SingleAndSequence['VSMapValueCallback[Any]']
+]
+
+BoundVSMapValue = TypeVar('BoundVSMapValue', bound=_VapourSynthMapValue)
+
+VSMapValueCallback = Callable[..., BoundVSMapValue]
+
+
+class _Future(Generic[T]):
+    def set_result(self, value: T) -> None: ...
+
+    def set_exception(self, exception: BaseException) -> None: ...
+
+    def result(self) -> T: ...
+
+    def exception(self) -> Union[NoReturn, None]: ...
+
+###
+# Typed dicts
+
+
+class _VideoFormatInfo(TypedDict):
+    id: int
+    name: str
+    color_family: 'ColorFamily'
+    sample_type: 'SampleType'
+    bits_per_sample: int
+    bytes_per_sample: int
+    subsampling_w: int
+    subsampling_h: int
+    num_planes: int
+
+
+###
+# VapourSynth Versioning
+
+
+class VapourSynthVersion(NamedTuple):
+    release_major: int
+    release_minor: int
+
+
+class VapourSynthAPIVersion(NamedTuple):
+    api_major: int
+    api_minor: int
+
+
+__version__: VapourSynthVersion
+__api_version__: VapourSynthAPIVersion
+
+###
+# VapourSynth Cython capsule
+
+
+class PyCapsule(Structure):
+    ...
+
+
+__pyx_capi__: Dict[str, PyCapsule]
+
+###
+# VapourSynth Enums and Constants
+
+
+class MessageType(IntEnum):
+    MESSAGE_TYPE_DEBUG: 'MessageType'
+    MESSAGE_TYPE_INFORMATION: 'MessageType'
+    MESSAGE_TYPE_WARNING: 'MessageType'
+    MESSAGE_TYPE_CRITICAL: 'MessageType'
+    MESSAGE_TYPE_FATAL: 'MessageType'
+
+
+MESSAGE_TYPE_DEBUG: MessageType
+MESSAGE_TYPE_INFORMATION: MessageType
+MESSAGE_TYPE_WARNING: MessageType
+MESSAGE_TYPE_CRITICAL: MessageType
+MESSAGE_TYPE_FATAL: MessageType
+
+
+class FilterMode(IntEnum):
+    fmParallel: 'FilterMode'
+    fmParallelRequests: 'FilterMode'
+    fmUnordered: 'FilterMode'
+    fmFrameState: 'FilterMode'
+
+
+fmParallel: FilterMode
+fmParallelRequests: FilterMode
+fmUnordered: FilterMode
+fmFrameState: FilterMode
+
+
+class CoreCreationFlags(IntEnum):
+    ccfEnableGraphInspection: 'CoreCreationFlags'
+    ccfDisableAutoLoading: 'CoreCreationFlags'
+    ccfDisableLibraryUnloading: 'CoreCreationFlags'
+
+
+ccfEnableGraphInspection: 'CoreCreationFlags'
+ccfDisableAutoLoading: 'CoreCreationFlags'
+ccfDisableLibraryUnloading: 'CoreCreationFlags'
+
+
+class MediaType(IntEnum):
+    VIDEO: 'MediaType'
+    AUDIO: 'MediaType'
+
+
+VIDEO: MediaType
+AUDIO: MediaType
+
+
+class ColorFamily(IntEnum):
+    UNDEFINED: 'ColorFamily'
+    GRAY: 'ColorFamily'
+    RGB: 'ColorFamily'
+    YUV: 'ColorFamily'
+
+
+UNDEFINED: ColorFamily
+GRAY: ColorFamily
+RGB: ColorFamily
+YUV: ColorFamily
+
+
+class ColorRange(IntEnum):
+    RANGE_FULL: 'ColorRange'
+    RANGE_LIMITED: 'ColorRange'
+
+
+RANGE_FULL: ColorRange
+RANGE_LIMITED: ColorRange
+
+
+class SampleType(IntEnum):
+    INTEGER: 'SampleType'
+    FLOAT: 'SampleType'
+
+
+INTEGER: SampleType
+FLOAT: SampleType
+
+
+class PresetFormat(IntEnum):
+    NONE: 'PresetFormat'
+
+    GRAY8: 'PresetFormat'
+    GRAY9: 'PresetFormat'
+    GRAY10: 'PresetFormat'
+    GRAY12: 'PresetFormat'
+    GRAY14: 'PresetFormat'
+    GRAY16: 'PresetFormat'
+    GRAY32: 'PresetFormat'
+
+    GRAYH: 'PresetFormat'
+    GRAYS: 'PresetFormat'
+
+    YUV420P8: 'PresetFormat'
+    YUV422P8: 'PresetFormat'
+    YUV444P8: 'PresetFormat'
+    YUV410P8: 'PresetFormat'
+    YUV411P8: 'PresetFormat'
+    YUV440P8: 'PresetFormat'
+
+    YUV420P9: 'PresetFormat'
+    YUV422P9: 'PresetFormat'
+    YUV444P9: 'PresetFormat'
+
+    YUV420P10: 'PresetFormat'
+    YUV422P10: 'PresetFormat'
+    YUV444P10: 'PresetFormat'
+
+    YUV420P12: 'PresetFormat'
+    YUV422P12: 'PresetFormat'
+    YUV444P12: 'PresetFormat'
+
+    YUV420P14: 'PresetFormat'
+    YUV422P14: 'PresetFormat'
+    YUV444P14: 'PresetFormat'
+
+    YUV420P16: 'PresetFormat'
+    YUV422P16: 'PresetFormat'
+    YUV444P16: 'PresetFormat'
+
+    YUV444PH: 'PresetFormat'
+    YUV444PS: 'PresetFormat'
+
+    RGB24: 'PresetFormat'
+    RGB27: 'PresetFormat'
+    RGB30: 'PresetFormat'
+    RGB36: 'PresetFormat'
+    RGB42: 'PresetFormat'
+    RGB48: 'PresetFormat'
+
+    RGBH: 'PresetFormat'
+    RGBS: 'PresetFormat'
+
+
+NONE: PresetFormat
+
+GRAY8: PresetFormat
+GRAY9: PresetFormat
+GRAY10: PresetFormat
+GRAY12: PresetFormat
+GRAY14: PresetFormat
+GRAY16: PresetFormat
+GRAY32: PresetFormat
+
+GRAYH: PresetFormat
+GRAYS: PresetFormat
+
+YUV420P8: PresetFormat
+YUV422P8: PresetFormat
+YUV444P8: PresetFormat
+YUV410P8: PresetFormat
+YUV411P8: PresetFormat
+YUV440P8: PresetFormat
+
+YUV420P9: PresetFormat
+YUV422P9: PresetFormat
+YUV444P9: PresetFormat
+
+YUV420P10: PresetFormat
+YUV422P10: PresetFormat
+YUV444P10: PresetFormat
+
+YUV420P12: PresetFormat
+YUV422P12: PresetFormat
+YUV444P12: PresetFormat
+
+YUV420P14: PresetFormat
+YUV422P14: PresetFormat
+YUV444P14: PresetFormat
+
+YUV420P16: PresetFormat
+YUV422P16: PresetFormat
+YUV444P16: PresetFormat
+
+YUV444PH: PresetFormat
+YUV444PS: PresetFormat
+
+RGB24: PresetFormat
+RGB27: PresetFormat
+RGB30: PresetFormat
+RGB36: PresetFormat
+RGB42: PresetFormat
+RGB48: PresetFormat
+
+RGBH: PresetFormat
+RGBS: PresetFormat
+
+
+class AudioChannels(IntEnum):
+    FRONT_LEFT: 'AudioChannels'
+    FRONT_RIGHT: 'AudioChannels'
+    FRONT_CENTER: 'AudioChannels'
+    LOW_FREQUENCY: 'AudioChannels'
+    BACK_LEFT: 'AudioChannels'
+    BACK_RIGHT: 'AudioChannels'
+    FRONT_LEFT_OF_CENTER: 'AudioChannels'
+    FRONT_RIGHT_OF_CENTER: 'AudioChannels'
+    BACK_CENTER: 'AudioChannels'
+    SIDE_LEFT: 'AudioChannels'
+    SIDE_RIGHT: 'AudioChannels'
+    TOP_CENTER: 'AudioChannels'
+    TOP_FRONT_LEFT: 'AudioChannels'
+    TOP_FRONT_CENTER: 'AudioChannels'
+    TOP_FRONT_RIGHT: 'AudioChannels'
+    TOP_BACK_LEFT: 'AudioChannels'
+    TOP_BACK_CENTER: 'AudioChannels'
+    TOP_BACK_RIGHT: 'AudioChannels'
+    STEREO_LEFT: 'AudioChannels'
+    STEREO_RIGHT: 'AudioChannels'
+    WIDE_LEFT: 'AudioChannels'
+    WIDE_RIGHT: 'AudioChannels'
+    SURROUND_DIRECT_LEFT: 'AudioChannels'
+    SURROUND_DIRECT_RIGHT: 'AudioChannels'
+    LOW_FREQUENCY2: 'AudioChannels'
+
+
+FRONT_LEFT: AudioChannels
+FRONT_RIGHT: AudioChannels
+FRONT_CENTER: AudioChannels
+LOW_FREQUENCY: AudioChannels
+BACK_LEFT: AudioChannels
+BACK_RIGHT: AudioChannels
+FRONT_LEFT_OF_CENTER: AudioChannels
+FRONT_RIGHT_OF_CENTER: AudioChannels
+BACK_CENTER: AudioChannels
+SIDE_LEFT: AudioChannels
+SIDE_RIGHT: AudioChannels
+TOP_CENTER: AudioChannels
+TOP_FRONT_LEFT: AudioChannels
+TOP_FRONT_CENTER: AudioChannels
+TOP_FRONT_RIGHT: AudioChannels
+TOP_BACK_LEFT: AudioChannels
+TOP_BACK_CENTER: AudioChannels
+TOP_BACK_RIGHT: AudioChannels
+STEREO_LEFT: AudioChannels
+STEREO_RIGHT: AudioChannels
+WIDE_LEFT: AudioChannels
+WIDE_RIGHT: AudioChannels
+SURROUND_DIRECT_LEFT: AudioChannels
+SURROUND_DIRECT_RIGHT: AudioChannels
+LOW_FREQUENCY2: AudioChannels
+
+
+class ChromaLocation(IntEnum):
+    CHROMA_LEFT: 'ChromaLocation'
+    CHROMA_CENTER: 'ChromaLocation'
+    CHROMA_TOP_LEFT: 'ChromaLocation'
+    CHROMA_TOP: 'ChromaLocation'
+    CHROMA_BOTTOM_LEFT: 'ChromaLocation'
+    CHROMA_BOTTOM: 'ChromaLocation'
+
+
+CHROMA_LEFT: ChromaLocation
+CHROMA_CENTER: ChromaLocation
+CHROMA_TOP_LEFT: ChromaLocation
+CHROMA_TOP: ChromaLocation
+CHROMA_BOTTOM_LEFT: ChromaLocation
+CHROMA_BOTTOM: ChromaLocation
+
+
+class FieldBased(IntEnum):
+    FIELD_PROGRESSIVE: 'FieldBased'
+    FIELD_TOP: 'FieldBased'
+    FIELD_BOTTOM: 'FieldBased'
+
+
+FIELD_PROGRESSIVE: FieldBased
+FIELD_TOP: FieldBased
+FIELD_BOTTOM: FieldBased
+
+
+class MatrixCoefficients(IntEnum):
+    MATRIX_RGB: 'MatrixCoefficients'
+    MATRIX_BT709: 'MatrixCoefficients'
+    MATRIX_UNSPECIFIED: 'MatrixCoefficients'
+    MATRIX_FCC: 'MatrixCoefficients'
+    MATRIX_BT470_BG: 'MatrixCoefficients'
+    MATRIX_ST170_M: 'MatrixCoefficients'
+    MATRIX_YCGCO: 'MatrixCoefficients'
+    MATRIX_BT2020_NCL: 'MatrixCoefficients'
+    MATRIX_BT2020_CL: 'MatrixCoefficients'
+    MATRIX_CHROMATICITY_DERIVED_NCL: 'MatrixCoefficients'
+    MATRIX_CHROMATICITY_DERIVED_CL: 'MatrixCoefficients'
+    MATRIX_ICTCP: 'MatrixCoefficients'
+
+
+MATRIX_RGB: MatrixCoefficients
+MATRIX_BT709: MatrixCoefficients
+MATRIX_UNSPECIFIED: MatrixCoefficients
+MATRIX_FCC: MatrixCoefficients
+MATRIX_BT470_BG: MatrixCoefficients
+MATRIX_ST170_M: MatrixCoefficients
+MATRIX_YCGCO: MatrixCoefficients
+MATRIX_BT2020_NCL: MatrixCoefficients
+MATRIX_BT2020_CL: MatrixCoefficients
+MATRIX_CHROMATICITY_DERIVED_NCL: MatrixCoefficients
+MATRIX_CHROMATICITY_DERIVED_CL: MatrixCoefficients
+MATRIX_ICTCP: MatrixCoefficients
+
+
+class TransferCharacteristics(IntEnum):
+    TRANSFER_BT709: 'TransferCharacteristics'
+    TRANSFER_UNSPECIFIED: 'TransferCharacteristics'
+    TRANSFER_BT470_M: 'TransferCharacteristics'
+    TRANSFER_BT470_BG: 'TransferCharacteristics'
+    TRANSFER_BT601: 'TransferCharacteristics'
+    TRANSFER_ST240_M: 'TransferCharacteristics'
+    TRANSFER_LINEAR: 'TransferCharacteristics'
+    TRANSFER_LOG_100: 'TransferCharacteristics'
+    TRANSFER_LOG_316: 'TransferCharacteristics'
+    TRANSFER_IEC_61966_2_4: 'TransferCharacteristics'
+    TRANSFER_IEC_61966_2_1: 'TransferCharacteristics'
+    TRANSFER_BT2020_10: 'TransferCharacteristics'
+    TRANSFER_BT2020_12: 'TransferCharacteristics'
+    TRANSFER_ST2084: 'TransferCharacteristics'
+    TRANSFER_ARIB_B67: 'TransferCharacteristics'
+
+
+TRANSFER_BT709: TransferCharacteristics
+TRANSFER_UNSPECIFIED: TransferCharacteristics
+TRANSFER_BT470_M: TransferCharacteristics
+TRANSFER_BT470_BG: TransferCharacteristics
+TRANSFER_BT601: TransferCharacteristics
+TRANSFER_ST240_M: TransferCharacteristics
+TRANSFER_LINEAR: TransferCharacteristics
+TRANSFER_LOG_100: TransferCharacteristics
+TRANSFER_LOG_316: TransferCharacteristics
+TRANSFER_IEC_61966_2_4: TransferCharacteristics
+TRANSFER_IEC_61966_2_1: TransferCharacteristics
+TRANSFER_BT2020_10: TransferCharacteristics
+TRANSFER_BT2020_12: TransferCharacteristics
+TRANSFER_ST2084: TransferCharacteristics
+TRANSFER_ARIB_B67: TransferCharacteristics
+
+
+class ColorPrimaries(IntEnum):
+    PRIMARIES_BT709: 'ColorPrimaries'
+    PRIMARIES_UNSPECIFIED: 'ColorPrimaries'
+    PRIMARIES_BT470_M: 'ColorPrimaries'
+    PRIMARIES_BT470_BG: 'ColorPrimaries'
+    PRIMARIES_ST170_M: 'ColorPrimaries'
+    PRIMARIES_ST240_M: 'ColorPrimaries'
+    PRIMARIES_FILM: 'ColorPrimaries'
+    PRIMARIES_BT2020: 'ColorPrimaries'
+    PRIMARIES_ST428: 'ColorPrimaries'
+    PRIMARIES_ST431_2: 'ColorPrimaries'
+    PRIMARIES_ST432_1: 'ColorPrimaries'
+    PRIMARIES_EBU3213_E: 'ColorPrimaries'
+
+
+PRIMARIES_BT709: ColorPrimaries
+PRIMARIES_UNSPECIFIED: ColorPrimaries
+PRIMARIES_BT470_M: ColorPrimaries
+PRIMARIES_BT470_BG: ColorPrimaries
+PRIMARIES_ST170_M: ColorPrimaries
+PRIMARIES_ST240_M: ColorPrimaries
+PRIMARIES_FILM: ColorPrimaries
+PRIMARIES_BT2020: ColorPrimaries
+PRIMARIES_ST428: ColorPrimaries
+PRIMARIES_ST431_2: ColorPrimaries
+PRIMARIES_ST432_1: ColorPrimaries
+PRIMARIES_EBU3213_E: ColorPrimaries
+
+
+###
+# VapourSynth Environment SubSystem
+
+
+class EnvironmentData:
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+
+class EnvironmentPolicy:
+    def on_policy_registered(self, special_api: 'EnvironmentPolicyAPI') -> None: ...
+
+    def on_policy_cleared(self) -> None: ...
+
+    @abstractmethod
+    def get_current_environment(self) -> Union[EnvironmentData, None]: ...
+
+    @abstractmethod
+    def set_environment(self, environment: Union[EnvironmentData, None]) -> Union[EnvironmentData, None]: ...
+
+    def is_alive(self, environment: EnvironmentData) -> bool: ...
+
+
+class StandaloneEnvironmentPolicy(EnvironmentPolicy):
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    def _on_log_message(self, level: MessageType, msg: str) -> None: ...
+
+    def on_policy_registered(self, api: 'EnvironmentPolicyAPI') -> None: ...
+
+    def on_policy_cleared(self) -> None: ...
+
+    def get_current_environment(self) -> EnvironmentData: ...
+
+    def set_environment(self, environment: Union[EnvironmentData, None]) -> EnvironmentData: ...
+
+    def is_alive(self, environment: EnvironmentData) -> bool: ...
+
+
+class VSScriptEnvironmentPolicy(EnvironmentPolicy):
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    def on_policy_registered(self, policy_api: 'EnvironmentPolicyAPI') -> None: ...
+
+    def on_policy_cleared(self) -> None: ...
+
+    def get_current_environment(self) -> Union[EnvironmentData, None]: ...
+
+    def set_environment(self, environment: Union[EnvironmentData, None]) -> Union[EnvironmentData, None]: ...
+
+    def is_alive(self, environment: EnvironmentData) -> bool: ...
+
+
+class EnvironmentPolicyAPI:
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    def wrap_environment(self, environment_data: EnvironmentData) -> 'Environment': ...
+
+    def create_environment(self, flags: int = 0) -> EnvironmentData: ...
+
+    def set_logger(self, env: EnvironmentData, logger: Callable[[int, str], None]) -> None: ...
+
+    def destroy_environment(self, env: EnvironmentData) -> None: ...
+
+    def unregister_policy(self) -> None: ...
+
+
+def register_policy(policy: EnvironmentPolicy) -> None:
+    ...
+
+
+if not TYPE_CHECKING:
+    def _try_enable_introspection(version: int = None): ...
+
+
+def has_policy() -> bool:
+    ...
+
+
+class Environment:
+    env: EnvironmentData
+
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    @property
+    def alive(self) -> bool: ...
+
+    @property
+    def single(self) -> bool: ...
+
+    @classmethod
+    def is_single(self) -> bool: ...
+
+    @property
+    def env_id(self) -> int: ...
+
+    @property
+    def active(self) -> bool: ...
+
+    def copy(self) -> 'Environment': ...
+
+    def use(self) -> ContextManager[None]: ...
+
+    def __eq__(self, other: 'Environment') -> bool: ...  # type: ignore[override]
+
+    def __repr__(self) -> str: ...
+
+
+def get_current_environment() -> Environment:
+    ...
+
+
+class VideoOutputTuple(NamedTuple):
+    clip: 'VideoNode'
+    alpha: Union['VideoNode', None]
+    alt_output: Literal[0, 1, 2]
+
+
+def _construct_type(signature: str) -> type:
+    ...
+
+
+def _construct_parameter(signature: str) -> Parameter:
+    ...
+
+
+def construct_signature(signature: str, return_signature: str, injected: Union[str, None] = None) -> Signature:
+    ...
+
+
+class Error(Exception):
+    ...
+
+
+def clear_output(index: int = 0) -> None:
+    ...
+
+
+def clear_outputs() -> None:
+    ...
+
+
+def get_outputs() -> MappingProxyType[int, Union[VideoOutputTuple, 'AudioNode']]:
+    ...
+
+
+def get_output(index: int = 0) -> Union[VideoOutputTuple, 'AudioNode']:
+    ...
+
+
+class FuncData:
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    def __call__(self, **kwargs: _VapourSynthMapValue) -> _VapourSynthMapValue: ...
+
+
+class Func:
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    def __call__(self, **kwargs: _VapourSynthMapValue) -> _VapourSynthMapValue: ...
+
+
+class FramePtr:
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+
+class VideoFormat:
+    id: int
+    name: str
+    color_family: ColorFamily
+    sample_type: SampleType
+    bits_per_sample: int
+    bytes_per_sample: int
+    subsampling_w: int
+    subsampling_h: int
+    num_planes: int
+
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    def _as_dict(self) -> _VideoFormatInfo: ...
+
+    def replace(
+        self, *,
+        color_family: Union[ColorFamily, None] = None,
+        sample_type: Union[SampleType, None] = None,
+        bits_per_sample: Union[int, None] = None,
+        subsampling_w: Union[int, None] = None,
+        subsampling_h: Union[int, None] = None
+    ) -> 'VideoFormat': ...
+
+    @overload
+    def __eq__(self, other: 'VideoFormat') -> bool: ...  # type: ignore[misc]
+
+    @overload
+    def __eq__(self, other: Any) -> Literal[False]: ...
+
+
+class FrameProps(MutableMapping[str, _VapourSynthMapValue]):
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    def setdefault(
+        self, key: str, default: _VapourSynthMapValue = 0
+    ) -> _VapourSynthMapValue: ...
+
+    def copy(self) -> 'FrameProps': ...
+
+    def __setattr__(self, name: str, value: _VapourSynthMapValue) -> None: ...
+
+    def __getattr__(self, name: str) -> _VapourSynthMapValue: ...
+
+    def __delattr__(self, name: str) -> None: ...
+
+    def __setitem__(self, name: str, value: _VapourSynthMapValue) -> None: ...
+
+    def __getitem__(self, name: str) -> _VapourSynthMapValue: ...
+
+    def __delitem__(self, name: str) -> None: ...
+
+    def __iter__(self) -> Iterator[str]: ...
+
+    def __len__(self) -> int: ...
+
+
+class RawFrame:
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    @property
+    def closed(self) -> bool: ...
+
+    def close(self) -> None: ...
+
+    def copy(self: 'SelfFrame') -> 'SelfFrame': ...
+
+    @property
+    def props(self) -> FrameProps: ...
+
+    @props.setter
+    def props(self, new_props: MappingProxyType[str, _VapourSynthMapValue]) -> None: ...
+
+    def get_write_ptr(self, plane: int) -> c_void_p: ...
+
+    def get_read_ptr(self, plane: int) -> c_void_p: ...
+
+    def get_stride(self, plane: int) -> int: ...
+
+    @property
+    def readonly(self) -> bool: ...
+
+    def __enter__(self: 'SelfFrame') -> 'SelfFrame': ...
+
+    def __exit__(
+        self, exc_type: Union[Type[BaseException], None],
+        exc_value: Union[BaseException, None],
+        traceback: Union[TracebackType, None], /,
+    ) -> Union[bool, None]: ...
+
+    def __getitem__(self, index: int) -> memoryview: ...
+
+    def __len__(self) -> int: ...
+
+
+SelfFrame = TypeVar('SelfFrame', bound=RawFrame)
+
+
+class VideoFrame(RawFrame):
+    format: VideoFormat
+    width: int
+    height: int
+
+    def _writelines(self, write: Callable[[bytes], None]) -> None: ...
+
+
+class AudioFrame(RawFrame):
+    sample_type: SampleType
+    bits_per_sample: int
+    bytes_per_sample: int
+    channel_layout: int
+    num_channels: int
+
+
+    
+# implementation: descale
+
+class _Plugin_descale_Core_Bound(Plugin):
+    """This class implements the module definitions for the "descale" VapourSynth plugin.\n\n*This class cannot be imported.*"""
+    def Debicubic(self, src: 'VideoNode', width: int, height: int, b: Optional[float] = None, c: Optional[float] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Debilinear(self, src: 'VideoNode', width: int, height: int, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Delanczos(self, src: 'VideoNode', width: int, height: int, taps: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Descale(self, src: 'VideoNode', width: int, height: int, kernel: Optional[DataType] = None, custom_kernel: Optional[VSMapValueCallback[Union[int, float]]] = None, taps: Optional[int] = None, b: Optional[float] = None, c: Optional[float] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Despline16(self, src: 'VideoNode', width: int, height: int, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Despline36(self, src: 'VideoNode', width: int, height: int, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Despline64(self, src: 'VideoNode', width: int, height: int, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+
+class _Plugin_descale_VideoNode_Bound(Plugin):
+    """This class implements the module definitions for the "descale" VapourSynth plugin.\n\n*This class cannot be imported.*"""
+    def Debicubic(self, width: int, height: int, b: Optional[float] = None, c: Optional[float] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Debilinear(self, width: int, height: int, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Delanczos(self, width: int, height: int, taps: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Descale(self, width: int, height: int, kernel: Optional[DataType] = None, custom_kernel: Optional[VSMapValueCallback[Union[int, float]]] = None, taps: Optional[int] = None, b: Optional[float] = None, c: Optional[float] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Despline16(self, width: int, height: int, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Despline36(self, width: int, height: int, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+    def Despline64(self, width: int, height: int, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, force: Optional[int] = None, force_h: Optional[int] = None, force_v: Optional[int] = None, opt: Optional[int] = None) -> 'VideoNode': ...
+
+# end implementation
+
+    
+# implementation: fmtc
+
+class _Plugin_fmtc_Core_Bound(Plugin):
+    """This class implements the module definitions for the "fmtc" VapourSynth plugin.\n\n*This class cannot be imported.*"""
+    def bitdepth(self, clip: 'VideoNode', csp: Optional[int] = None, bits: Optional[int] = None, flt: Optional[int] = None, planes: Optional[SingleAndSequence[int]] = None, fulls: Optional[int] = None, fulld: Optional[int] = None, dmode: Optional[int] = None, ampo: Optional[float] = None, ampn: Optional[float] = None, dyn: Optional[int] = None, staticnoise: Optional[int] = None, cpuopt: Optional[int] = None, patsize: Optional[int] = None, tpdfo: Optional[int] = None, tpdfn: Optional[int] = None, corplane: Optional[int] = None) -> 'VideoNode': ...
+    def histluma(self, clip: 'VideoNode', full: Optional[int] = None, amp: Optional[int] = None) -> 'VideoNode': ...
+    def matrix(self, clip: 'VideoNode', mat: Optional[DataType] = None, mats: Optional[DataType] = None, matd: Optional[DataType] = None, fulls: Optional[int] = None, fulld: Optional[int] = None, coef: Optional[SingleAndSequence[float]] = None, csp: Optional[int] = None, col_fam: Optional[int] = None, bits: Optional[int] = None, singleout: Optional[int] = None, cpuopt: Optional[int] = None, planes: Optional[SingleAndSequence[float]] = None) -> 'VideoNode': ...
+    def matrix2020cl(self, clip: 'VideoNode', full: Optional[int] = None, csp: Optional[int] = None, bits: Optional[int] = None, cpuopt: Optional[int] = None) -> 'VideoNode': ...
+    def nativetostack16(self, clip: 'VideoNode') -> 'VideoNode': ...
+    def primaries(self, clip: 'VideoNode', rs: Optional[SingleAndSequence[float]] = None, gs: Optional[SingleAndSequence[float]] = None, bs: Optional[SingleAndSequence[float]] = None, ws: Optional[SingleAndSequence[float]] = None, rd: Optional[SingleAndSequence[float]] = None, gd: Optional[SingleAndSequence[float]] = None, bd: Optional[SingleAndSequence[float]] = None, wd: Optional[SingleAndSequence[float]] = None, prims: Optional[DataType] = None, primd: Optional[DataType] = None, cpuopt: Optional[int] = None) -> 'VideoNode': ...
+    def resample(self, clip: 'VideoNode', w: Optional[int] = None, h: Optional[int] = None, sx: Optional[SingleAndSequence[float]] = None, sy: Optional[SingleAndSequence[float]] = None, sw: Optional[SingleAndSequence[float]] = None, sh: Optional[SingleAndSequence[float]] = None, scale: Optional[float] = None, scaleh: Optional[float] = None, scalev: Optional[float] = None, kernel: Optional[SingleAndSequence[DataType]] = None, kernelh: Optional[SingleAndSequence[DataType]] = None, kernelv: Optional[SingleAndSequence[DataType]] = None, impulse: Optional[SingleAndSequence[float]] = None, impulseh: Optional[SingleAndSequence[float]] = None, impulsev: Optional[SingleAndSequence[float]] = None, taps: Optional[SingleAndSequence[int]] = None, tapsh: Optional[SingleAndSequence[int]] = None, tapsv: Optional[SingleAndSequence[int]] = None, a1: Optional[SingleAndSequence[float]] = None, a2: Optional[SingleAndSequence[float]] = None, a3: Optional[SingleAndSequence[float]] = None, a1h: Optional[SingleAndSequence[float]] = None, a2h: Optional[SingleAndSequence[float]] = None, a3h: Optional[SingleAndSequence[float]] = None, a1v: Optional[SingleAndSequence[float]] = None, a2v: Optional[SingleAndSequence[float]] = None, a3v: Optional[SingleAndSequence[float]] = None, kovrspl: Optional[SingleAndSequence[int]] = None, fh: Optional[SingleAndSequence[float]] = None, fv: Optional[SingleAndSequence[float]] = None, cnorm: Optional[SingleAndSequence[int]] = None, total: Optional[SingleAndSequence[float]] = None, totalh: Optional[SingleAndSequence[float]] = None, totalv: Optional[SingleAndSequence[float]] = None, invks: Optional[SingleAndSequence[int]] = None, invksh: Optional[SingleAndSequence[int]] = None, invksv: Optional[SingleAndSequence[int]] = None, invkstaps: Optional[SingleAndSequence[int]] = None, invkstapsh: Optional[SingleAndSequence[int]] = None, invkstapsv: Optional[SingleAndSequence[int]] = None, csp: Optional[int] = None, css: Optional[DataType] = None, planes: Optional[SingleAndSequence[float]] = None, fulls: Optional[int] = None, fulld: Optional[int] = None, center: Optional[SingleAndSequence[int]] = None, cplace: Optional[DataType] = None, cplaces: Optional[DataType] = None, cplaced: Optional[DataType] = None, interlaced: Optional[int] = None, interlacedd: Optional[int] = None, tff: Optional[int] = None, tffd: Optional[int] = None, flt: Optional[int] = None, cpuopt: Optional[int] = None) -> 'VideoNode': ...
+    def stack16tonative(self, clip: 'VideoNode') -> 'VideoNode': ...
+    def transfer(self, clip: 'VideoNode', transs: Optional[SingleAndSequence[DataType]] = None, transd: Optional[SingleAndSequence[DataType]] = None, cont: Optional[float] = None, gcor: Optional[float] = None, bits: Optional[int] = None, flt: Optional[int] = None, fulls: Optional[int] = None, fulld: Optional[int] = None, logceis: Optional[int] = None, logceid: Optional[int] = None, cpuopt: Optional[int] = None, blacklvl: Optional[float] = None, sceneref: Optional[int] = None, lb: Optional[float] = None, lw: Optional[float] = None, lws: Optional[float] = None, lwd: Optional[float] = None, ambient: Optional[float] = None, match: Optional[int] = None, gy: Optional[int] = None, debug: Optional[int] = None) -> 'VideoNode': ...
+
+class _Plugin_fmtc_VideoNode_Bound(Plugin):
+    """This class implements the module definitions for the "fmtc" VapourSynth plugin.\n\n*This class cannot be imported.*"""
+    def bitdepth(self, csp: Optional[int] = None, bits: Optional[int] = None, flt: Optional[int] = None, planes: Optional[SingleAndSequence[int]] = None, fulls: Optional[int] = None, fulld: Optional[int] = None, dmode: Optional[int] = None, ampo: Optional[float] = None, ampn: Optional[float] = None, dyn: Optional[int] = None, staticnoise: Optional[int] = None, cpuopt: Optional[int] = None, patsize: Optional[int] = None, tpdfo: Optional[int] = None, tpdfn: Optional[int] = None, corplane: Optional[int] = None) -> 'VideoNode': ...
+    def histluma(self, full: Optional[int] = None, amp: Optional[int] = None) -> 'VideoNode': ...
+    def matrix(self, mat: Optional[DataType] = None, mats: Optional[DataType] = None, matd: Optional[DataType] = None, fulls: Optional[int] = None, fulld: Optional[int] = None, coef: Optional[SingleAndSequence[float]] = None, csp: Optional[int] = None, col_fam: Optional[int] = None, bits: Optional[int] = None, singleout: Optional[int] = None, cpuopt: Optional[int] = None, planes: Optional[SingleAndSequence[float]] = None) -> 'VideoNode': ...
+    def matrix2020cl(self, full: Optional[int] = None, csp: Optional[int] = None, bits: Optional[int] = None, cpuopt: Optional[int] = None) -> 'VideoNode': ...
+    def nativetostack16(self) -> 'VideoNode': ...
+    def primaries(self, rs: Optional[SingleAndSequence[float]] = None, gs: Optional[SingleAndSequence[float]] = None, bs: Optional[SingleAndSequence[float]] = None, ws: Optional[SingleAndSequence[float]] = None, rd: Optional[SingleAndSequence[float]] = None, gd: Optional[SingleAndSequence[float]] = None, bd: Optional[SingleAndSequence[float]] = None, wd: Optional[SingleAndSequence[float]] = None, prims: Optional[DataType] = None, primd: Optional[DataType] = None, cpuopt: Optional[int] = None) -> 'VideoNode': ...
+    def resample(self, w: Optional[int] = None, h: Optional[int] = None, sx: Optional[SingleAndSequence[float]] = None, sy: Optional[SingleAndSequence[float]] = None, sw: Optional[SingleAndSequence[float]] = None, sh: Optional[SingleAndSequence[float]] = None, scale: Optional[float] = None, scaleh: Optional[float] = None, scalev: Optional[float] = None, kernel: Optional[SingleAndSequence[DataType]] = None, kernelh: Optional[SingleAndSequence[DataType]] = None, kernelv: Optional[SingleAndSequence[DataType]] = None, impulse: Optional[SingleAndSequence[float]] = None, impulseh: Optional[SingleAndSequence[float]] = None, impulsev: Optional[SingleAndSequence[float]] = None, taps: Optional[SingleAndSequence[int]] = None, tapsh: Optional[SingleAndSequence[int]] = None, tapsv: Optional[SingleAndSequence[int]] = None, a1: Optional[SingleAndSequence[float]] = None, a2: Optional[SingleAndSequence[float]] = None, a3: Optional[SingleAndSequence[float]] = None, a1h: Optional[SingleAndSequence[float]] = None, a2h: Optional[SingleAndSequence[float]] = None, a3h: Optional[SingleAndSequence[float]] = None, a1v: Optional[SingleAndSequence[float]] = None, a2v: Optional[SingleAndSequence[float]] = None, a3v: Optional[SingleAndSequence[float]] = None, kovrspl: Optional[SingleAndSequence[int]] = None, fh: Optional[SingleAndSequence[float]] = None, fv: Optional[SingleAndSequence[float]] = None, cnorm: Optional[SingleAndSequence[int]] = None, total: Optional[SingleAndSequence[float]] = None, totalh: Optional[SingleAndSequence[float]] = None, totalv: Optional[SingleAndSequence[float]] = None, invks: Optional[SingleAndSequence[int]] = None, invksh: Optional[SingleAndSequence[int]] = None, invksv: Optional[SingleAndSequence[int]] = None, invkstaps: Optional[SingleAndSequence[int]] = None, invkstapsh: Optional[SingleAndSequence[int]] = None, invkstapsv: Optional[SingleAndSequence[int]] = None, csp: Optional[int] = None, css: Optional[DataType] = None, planes: Optional[SingleAndSequence[float]] = None, fulls: Optional[int] = None, fulld: Optional[int] = None, center: Optional[SingleAndSequence[int]] = None, cplace: Optional[DataType] = None, cplaces: Optional[DataType] = None, cplaced: Optional[DataType] = None, interlaced: Optional[int] = None, interlacedd: Optional[int] = None, tff: Optional[int] = None, tffd: Optional[int] = None, flt: Optional[int] = None, cpuopt: Optional[int] = None) -> 'VideoNode': ...
+    def stack16tonative(self) -> 'VideoNode': ...
+    def transfer(self, transs: Optional[SingleAndSequence[DataType]] = None, transd: Optional[SingleAndSequence[DataType]] = None, cont: Optional[float] = None, gcor: Optional[float] = None, bits: Optional[int] = None, flt: Optional[int] = None, fulls: Optional[int] = None, fulld: Optional[int] = None, logceis: Optional[int] = None, logceid: Optional[int] = None, cpuopt: Optional[int] = None, blacklvl: Optional[float] = None, sceneref: Optional[int] = None, lb: Optional[float] = None, lw: Optional[float] = None, lws: Optional[float] = None, lwd: Optional[float] = None, ambient: Optional[float] = None, match: Optional[int] = None, gy: Optional[int] = None, debug: Optional[int] = None) -> 'VideoNode': ...
+
+# end implementation
+
+    
+# implementation: placebo
+
+class _Plugin_placebo_Core_Bound(Plugin):
+    """This class implements the module definitions for the "placebo" VapourSynth plugin.\n\n*This class cannot be imported.*"""
+    def Deband(self, clip: 'VideoNode', planes: Optional[int] = None, iterations: Optional[int] = None, threshold: Optional[float] = None, radius: Optional[float] = None, grain: Optional[float] = None, dither: Optional[int] = None, dither_algo: Optional[int] = None, renderer_api: Optional[int] = None) -> 'VideoNode': ...
+    def Resample(self, clip: 'VideoNode', width: int, height: int, filter: Optional[DataType] = None, clamp: Optional[float] = None, blur: Optional[float] = None, taper: Optional[float] = None, radius: Optional[float] = None, param1: Optional[float] = None, param2: Optional[float] = None, sx: Optional[float] = None, sy: Optional[float] = None, antiring: Optional[float] = None, lut_entries: Optional[int] = None, cutoff: Optional[float] = None, sigmoidize: Optional[int] = None, sigmoid_center: Optional[float] = None, sigmoid_slope: Optional[float] = None, linearize: Optional[int] = None, trc: Optional[int] = None) -> 'VideoNode': ...
+    def Shader(self, clip: 'VideoNode', shader: DataType, width: Optional[int] = None, height: Optional[int] = None, chroma_loc: Optional[int] = None, matrix: Optional[int] = None, trc: Optional[int] = None, linearize: Optional[int] = None, sigmoidize: Optional[int] = None, sigmoid_center: Optional[float] = None, sigmoid_slope: Optional[float] = None, lut_entries: Optional[int] = None, antiring: Optional[float] = None, filter: Optional[DataType] = None, clamp: Optional[float] = None, blur: Optional[float] = None, taper: Optional[float] = None, radius: Optional[float] = None, param1: Optional[float] = None, param2: Optional[float] = None) -> 'VideoNode': ...
+    def Tonemap(self, clip: 'VideoNode', srcp: Optional[int] = None, srct: Optional[int] = None, srcl: Optional[int] = None, src_peak: Optional[float] = None, src_avg: Optional[float] = None, src_scale: Optional[float] = None, dstp: Optional[int] = None, dstt: Optional[int] = None, dstl: Optional[int] = None, dst_peak: Optional[float] = None, dst_avg: Optional[float] = None, dst_scale: Optional[float] = None, dynamic_peak_detection: Optional[int] = None, smoothing_period: Optional[float] = None, scene_threshold_low: Optional[float] = None, scene_threshold_high: Optional[float] = None, intent: Optional[int] = None, tone_mapping_algo: Optional[int] = None, tone_mapping_param: Optional[float] = None, desaturation_strength: Optional[float] = None, desaturation_exponent: Optional[float] = None, desaturation_base: Optional[float] = None, max_boost: Optional[float] = None, gamut_warning: Optional[int] = None) -> 'VideoNode': ...
+
+class _Plugin_placebo_VideoNode_Bound(Plugin):
+    """This class implements the module definitions for the "placebo" VapourSynth plugin.\n\n*This class cannot be imported.*"""
+    def Deband(self, planes: Optional[int] = None, iterations: Optional[int] = None, threshold: Optional[float] = None, radius: Optional[float] = None, grain: Optional[float] = None, dither: Optional[int] = None, dither_algo: Optional[int] = None, renderer_api: Optional[int] = None) -> 'VideoNode': ...
+    def Resample(self, width: int, height: int, filter: Optional[DataType] = None, clamp: Optional[float] = None, blur: Optional[float] = None, taper: Optional[float] = None, radius: Optional[float] = None, param1: Optional[float] = None, param2: Optional[float] = None, sx: Optional[float] = None, sy: Optional[float] = None, antiring: Optional[float] = None, lut_entries: Optional[int] = None, cutoff: Optional[float] = None, sigmoidize: Optional[int] = None, sigmoid_center: Optional[float] = None, sigmoid_slope: Optional[float] = None, linearize: Optional[int] = None, trc: Optional[int] = None) -> 'VideoNode': ...
+    def Shader(self, shader: DataType, width: Optional[int] = None, height: Optional[int] = None, chroma_loc: Optional[int] = None, matrix: Optional[int] = None, trc: Optional[int] = None, linearize: Optional[int] = None, sigmoidize: Optional[int] = None, sigmoid_center: Optional[float] = None, sigmoid_slope: Optional[float] = None, lut_entries: Optional[int] = None, antiring: Optional[float] = None, filter: Optional[DataType] = None, clamp: Optional[float] = None, blur: Optional[float] = None, taper: Optional[float] = None, radius: Optional[float] = None, param1: Optional[float] = None, param2: Optional[float] = None) -> 'VideoNode': ...
+    def Tonemap(self, srcp: Optional[int] = None, srct: Optional[int] = None, srcl: Optional[int] = None, src_peak: Optional[float] = None, src_avg: Optional[float] = None, src_scale: Optional[float] = None, dstp: Optional[int] = None, dstt: Optional[int] = None, dstl: Optional[int] = None, dst_peak: Optional[float] = None, dst_avg: Optional[float] = None, dst_scale: Optional[float] = None, dynamic_peak_detection: Optional[int] = None, smoothing_period: Optional[float] = None, scene_threshold_low: Optional[float] = None, scene_threshold_high: Optional[float] = None, intent: Optional[int] = None, tone_mapping_algo: Optional[int] = None, tone_mapping_param: Optional[float] = None, desaturation_strength: Optional[float] = None, desaturation_exponent: Optional[float] = None, desaturation_base: Optional[float] = None, max_boost: Optional[float] = None, gamut_warning: Optional[int] = None) -> 'VideoNode': ...
+
+# end implementation
+
+    
+# implementation: resize
+
+class _Plugin_resize_Core_Bound(Plugin):
+    """This class implements the module definitions for the "resize" VapourSynth plugin.\n\n*This class cannot be imported.*"""
+    def Bicubic(self, clip: 'VideoNode', width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Bilinear(self, clip: 'VideoNode', width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Bob(self, clip: 'VideoNode', filter: Optional[DataType] = None, tff: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Lanczos(self, clip: 'VideoNode', width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Point(self, clip: 'VideoNode', width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Spline16(self, clip: 'VideoNode', width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Spline36(self, clip: 'VideoNode', width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Spline64(self, clip: 'VideoNode', width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+
+class _Plugin_resize_VideoNode_Bound(Plugin):
+    """This class implements the module definitions for the "resize" VapourSynth plugin.\n\n*This class cannot be imported.*"""
+    def Bicubic(self, width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Bilinear(self, width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Bob(self, filter: Optional[DataType] = None, tff: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Lanczos(self, width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Point(self, width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Spline16(self, width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Spline36(self, width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+    def Spline64(self, width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, matrix: Optional[int] = None, matrix_s: Optional[DataType] = None, transfer: Optional[int] = None, transfer_s: Optional[DataType] = None, primaries: Optional[int] = None, primaries_s: Optional[DataType] = None, range: Optional[int] = None, range_s: Optional[DataType] = None, chromaloc: Optional[int] = None, chromaloc_s: Optional[DataType] = None, matrix_in: Optional[int] = None, matrix_in_s: Optional[DataType] = None, transfer_in: Optional[int] = None, transfer_in_s: Optional[DataType] = None, primaries_in: Optional[int] = None, primaries_in_s: Optional[DataType] = None, range_in: Optional[int] = None, range_in_s: Optional[DataType] = None, chromaloc_in: Optional[int] = None, chromaloc_in_s: Optional[DataType] = None, filter_param_a: Optional[float] = None, filter_param_b: Optional[float] = None, resample_filter_uv: Optional[DataType] = None, filter_param_a_uv: Optional[float] = None, filter_param_b_uv: Optional[float] = None, dither_type: Optional[DataType] = None, cpu_type: Optional[DataType] = None, prefer_props: Optional[int] = None, src_left: Optional[float] = None, src_top: Optional[float] = None, src_width: Optional[float] = None, src_height: Optional[float] = None, nominal_luminance: Optional[float] = None) -> 'VideoNode': ...
+
+# end implementation
+
+    
+# implementation: std
+
+class _Plugin_std_Core_Bound(Plugin):
+    """This class implements the module definitions for the "std" VapourSynth plugin.\n\n*This class cannot be imported.*"""
+    def AddBorders(self, clip: 'VideoNode', left: Optional[int] = None, right: Optional[int] = None, top: Optional[int] = None, bottom: Optional[int] = None, color: Optional[SingleAndSequence[float]] = None) -> 'VideoNode': ...
+    def AssumeFPS(self, clip: 'VideoNode', src: Optional['VideoNode'] = None, fpsnum: Optional[int] = None, fpsden: Optional[int] = None) -> 'VideoNode': ...
+    def AssumeSampleRate(self, clip: 'AudioNode', src: Optional['AudioNode'] = None, samplerate: Optional[int] = None) -> 'AudioNode': ...
+    def AudioGain(self, clip: 'AudioNode', gain: Optional[SingleAndSequence[float]] = None) -> 'AudioNode': ...
+    def AudioLoop(self, clip: 'AudioNode', times: Optional[int] = None) -> 'AudioNode': ...
+    def AudioMix(self, clips: SingleAndSequence['AudioNode'], matrix: SingleAndSequence[float], channels_out: SingleAndSequence[int]) -> 'AudioNode': ...
+    def AudioReverse(self, clip: 'AudioNode') -> 'AudioNode': ...
+    def AudioSplice(self, clips: SingleAndSequence['AudioNode']) -> 'AudioNode': ...
+    def AudioTrim(self, clip: 'AudioNode', first: Optional[int] = None, last: Optional[int] = None, length: Optional[int] = None) -> 'AudioNode': ...
+    def AverageFrames(self, clips: SingleAndSequence['VideoNode'], weights: SingleAndSequence[float], scale: Optional[float] = None, scenechange: Optional[int] = None, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Binarize(self, clip: 'VideoNode', threshold: Optional[SingleAndSequence[float]] = None, v0: Optional[SingleAndSequence[float]] = None, v1: Optional[SingleAndSequence[float]] = None, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def BinarizeMask(self, clip: 'VideoNode', threshold: Optional[SingleAndSequence[float]] = None, v0: Optional[SingleAndSequence[float]] = None, v1: Optional[SingleAndSequence[float]] = None, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def BlankAudio(self, clip: Optional['AudioNode'] = None, channels: Optional[int] = None, bits: Optional[int] = None, sampletype: Optional[int] = None, samplerate: Optional[int] = None, length: Optional[int] = None, keep: Optional[int] = None) -> 'AudioNode': ...
+    def BlankClip(self, clip: Optional['VideoNode'] = None, width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, length: Optional[int] = None, fpsnum: Optional[int] = None, fpsden: Optional[int] = None, color: Optional[SingleAndSequence[float]] = None, keep: Optional[int] = None) -> 'VideoNode': ...
+    def BoxBlur(self, clip: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, hradius: Optional[int] = None, hpasses: Optional[int] = None, vradius: Optional[int] = None, vpasses: Optional[int] = None) -> 'VideoNode': ...
+    def Cache(self, clip: 'VideoNode', size: Optional[int] = None, fixed: Optional[int] = None, make_linear: Optional[int] = None) -> 'VideoNode': ...
+    def ClipToProp(self, clip: 'VideoNode', mclip: 'VideoNode', prop: Optional[DataType] = None) -> 'VideoNode': ...
+    def Convolution(self, clip: 'VideoNode', matrix: SingleAndSequence[float], bias: Optional[float] = None, divisor: Optional[float] = None, planes: Optional[SingleAndSequence[int]] = None, saturate: Optional[int] = None, mode: Optional[DataType] = None) -> 'VideoNode': ...
+    def CopyFrameProps(self, clip: 'VideoNode', prop_src: 'VideoNode') -> 'VideoNode': ...
+    def Crop(self, clip: 'VideoNode', left: Optional[int] = None, right: Optional[int] = None, top: Optional[int] = None, bottom: Optional[int] = None) -> 'VideoNode': ...
+    def CropAbs(self, clip: 'VideoNode', width: int, height: int, left: Optional[int] = None, top: Optional[int] = None, x: Optional[int] = None, y: Optional[int] = None) -> 'VideoNode': ...
+    def CropRel(self, clip: 'VideoNode', left: Optional[int] = None, right: Optional[int] = None, top: Optional[int] = None, bottom: Optional[int] = None) -> 'VideoNode': ...
+    def Deflate(self, clip: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, threshold: Optional[float] = None) -> 'VideoNode': ...
+    def DeleteFrames(self, clip: 'VideoNode', frames: SingleAndSequence[int]) -> 'VideoNode': ...
+    def DoubleWeave(self, clip: 'VideoNode', tff: Optional[int] = None) -> 'VideoNode': ...
+    def DuplicateFrames(self, clip: 'VideoNode', frames: SingleAndSequence[int]) -> 'VideoNode': ...
+    def Expr(self, clips: SingleAndSequence['VideoNode'], expr: SingleAndSequence[DataType], format: Optional[int] = None) -> 'VideoNode': ...
+    def FlipHorizontal(self, clip: 'VideoNode') -> 'VideoNode': ...
+    def FlipVertical(self, clip: 'VideoNode') -> 'VideoNode': ...
+    def FrameEval(self, clip: 'VideoNode', eval: VSMapValueCallback[VideoNode], prop_src: Optional[SingleAndSequence[VideoNode]] = None, clip_src: Optional[SingleAndSequence[VideoNode]] = None) -> 'VideoNode': ...
+    def FreezeFrames(self, clip: 'VideoNode', first: SingleAndSequence[int], last: SingleAndSequence[int], replacement: SingleAndSequence[int]) -> 'VideoNode': ...
+    def Inflate(self, clip: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, threshold: Optional[float] = None) -> 'VideoNode': ...
+    def Interleave(self, clips: SingleAndSequence['VideoNode'], extend: Optional[int] = None, mismatch: Optional[int] = None, modify_duration: Optional[int] = None) -> 'VideoNode': ...
+    def Invert(self, clip: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def InvertMask(self, clip: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Levels(self, clip: 'VideoNode', min_in: Optional[SingleAndSequence[float]] = None, max_in: Optional[SingleAndSequence[float]] = None, gamma: Optional[SingleAndSequence[float]] = None, min_out: Optional[SingleAndSequence[float]] = None, max_out: Optional[SingleAndSequence[float]] = None, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Limiter(self, clip: 'VideoNode', min: Optional[SingleAndSequence[float]] = None, max: Optional[SingleAndSequence[float]] = None, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def LoadAllPlugins(self, path: DataType) -> None: ...
+    def LoadPlugin(self, path: DataType, altsearchpath: Optional[int] = None, forcens: Optional[DataType] = None, forceid: Optional[DataType] = None) -> None: ...
+    def Loop(self, clip: 'VideoNode', times: Optional[int] = None) -> 'VideoNode': ...
+    def Lut(self, clip: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, lut: Optional[SingleAndSequence[int]] = None, lutf: Optional[SingleAndSequence[float]] = None, function: Optional[VSMapValueCallback[Union[int, float]]] = None, bits: Optional[int] = None, floatout: Optional[int] = None) -> 'VideoNode': ...
+    def Lut2(self, clipa: 'VideoNode', clipb: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, lut: Optional[SingleAndSequence[int]] = None, lutf: Optional[SingleAndSequence[float]] = None, function: Optional[VSMapValueCallback[Union[int, float]]] = None, bits: Optional[int] = None, floatout: Optional[int] = None) -> 'VideoNode': ...
+    def MakeDiff(self, clipa: 'VideoNode', clipb: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def MaskedMerge(self, clipa: 'VideoNode', clipb: 'VideoNode', mask: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, first_plane: Optional[int] = None, premultiplied: Optional[int] = None) -> 'VideoNode': ...
+    def Maximum(self, clip: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, threshold: Optional[float] = None, coordinates: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Median(self, clip: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Merge(self, clipa: 'VideoNode', clipb: 'VideoNode', weight: Optional[SingleAndSequence[float]] = None) -> 'VideoNode': ...
+    def MergeDiff(self, clipa: 'VideoNode', clipb: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Minimum(self, clip: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, threshold: Optional[float] = None, coordinates: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def ModifyFrame(self, clip: 'VideoNode', clips: SingleAndSequence['VideoNode'], selector: VSMapValueCallback[VideoFrame]) -> 'VideoNode': ...
+    def PEMVerifier(self, clip: 'VideoNode', upper: Optional[SingleAndSequence[float]] = None, lower: Optional[SingleAndSequence[float]] = None) -> 'VideoNode': ...
+    def PlaneStats(self, clipa: 'VideoNode', clipb: Optional['VideoNode'] = None, plane: Optional[int] = None, prop: Optional[DataType] = None) -> 'VideoNode': ...
+    def PreMultiply(self, clip: 'VideoNode', alpha: 'VideoNode') -> 'VideoNode': ...
+    def Prewitt(self, clip: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, scale: Optional[float] = None) -> 'VideoNode': ...
+    def PropToClip(self, clip: 'VideoNode', prop: Optional[DataType] = None) -> 'VideoNode': ...
+    def RemoveFrameProps(self, clip: 'VideoNode', props: Optional[SingleAndSequence[DataType]] = None) -> 'VideoNode': ...
+    def Reverse(self, clip: 'VideoNode') -> 'VideoNode': ...
+    def SelectEvery(self, clip: 'VideoNode', cycle: int, offsets: SingleAndSequence[int], modify_duration: Optional[int] = None) -> 'VideoNode': ...
+    def SeparateFields(self, clip: 'VideoNode', tff: Optional[int] = None, modify_duration: Optional[int] = None) -> 'VideoNode': ...
+    def SetAudioCache(self, clip: 'AudioNode', mode: Optional[int] = None, fixedsize: Optional[int] = None, maxsize: Optional[int] = None, maxhistory: Optional[int] = None) -> None: ...
+    def SetFieldBased(self, clip: 'VideoNode', value: int) -> 'VideoNode': ...
+    def SetFrameProp(self, clip: 'VideoNode', prop: DataType, intval: Optional[SingleAndSequence[int]] = None, floatval: Optional[SingleAndSequence[float]] = None, data: Optional[SingleAndSequence[DataType]] = None) -> 'VideoNode': ...
+    def SetFrameProps(self, clip: 'VideoNode', **kwargs: _VapourSynthMapValue) -> 'VideoNode': ...
+    def SetMaxCPU(self, cpu: DataType) -> DataType: ...
+    def SetVideoCache(self, clip: 'VideoNode', mode: Optional[int] = None, fixedsize: Optional[int] = None, maxsize: Optional[int] = None, maxhistory: Optional[int] = None) -> None: ...
+    def ShuffleChannels(self, clips: SingleAndSequence['AudioNode'], channels_in: SingleAndSequence[int], channels_out: SingleAndSequence[int]) -> 'AudioNode': ...
+    def ShufflePlanes(self, clips: SingleAndSequence['VideoNode'], planes: SingleAndSequence[int], colorfamily: int) -> 'VideoNode': ...
+    def Sobel(self, clip: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, scale: Optional[float] = None) -> 'VideoNode': ...
+    def Splice(self, clips: SingleAndSequence['VideoNode'], mismatch: Optional[int] = None) -> 'VideoNode': ...
+    def SplitChannels(self, clip: 'AudioNode') -> SingleAndSequence['AudioNode']: ...
+    def SplitPlanes(self, clip: 'VideoNode') -> SingleAndSequence['VideoNode']: ...
+    def StackHorizontal(self, clips: SingleAndSequence['VideoNode']) -> 'VideoNode': ...
+    def StackVertical(self, clips: SingleAndSequence['VideoNode']) -> 'VideoNode': ...
+    def TestAudio(self, channels: Optional[int] = None, bits: Optional[int] = None, isfloat: Optional[int] = None, samplerate: Optional[int] = None, length: Optional[int] = None) -> 'AudioNode': ...
+    def Transpose(self, clip: 'VideoNode') -> 'VideoNode': ...
+    def Trim(self, clip: 'VideoNode', first: Optional[int] = None, last: Optional[int] = None, length: Optional[int] = None) -> 'VideoNode': ...
+    def Turn180(self, clip: 'VideoNode') -> 'VideoNode': ...
+
+class _Plugin_std_VideoNode_Bound(Plugin):
+    """This class implements the module definitions for the "std" VapourSynth plugin.\n\n*This class cannot be imported.*"""
+    def AddBorders(self, left: Optional[int] = None, right: Optional[int] = None, top: Optional[int] = None, bottom: Optional[int] = None, color: Optional[SingleAndSequence[float]] = None) -> 'VideoNode': ...
+    def AssumeFPS(self, src: Optional['VideoNode'] = None, fpsnum: Optional[int] = None, fpsden: Optional[int] = None) -> 'VideoNode': ...
+    def AverageFrames(self, weights: SingleAndSequence[float], scale: Optional[float] = None, scenechange: Optional[int] = None, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Binarize(self, threshold: Optional[SingleAndSequence[float]] = None, v0: Optional[SingleAndSequence[float]] = None, v1: Optional[SingleAndSequence[float]] = None, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def BinarizeMask(self, threshold: Optional[SingleAndSequence[float]] = None, v0: Optional[SingleAndSequence[float]] = None, v1: Optional[SingleAndSequence[float]] = None, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def BlankClip(self, width: Optional[int] = None, height: Optional[int] = None, format: Optional[int] = None, length: Optional[int] = None, fpsnum: Optional[int] = None, fpsden: Optional[int] = None, color: Optional[SingleAndSequence[float]] = None, keep: Optional[int] = None) -> 'VideoNode': ...
+    def BoxBlur(self, planes: Optional[SingleAndSequence[int]] = None, hradius: Optional[int] = None, hpasses: Optional[int] = None, vradius: Optional[int] = None, vpasses: Optional[int] = None) -> 'VideoNode': ...
+    def Cache(self, size: Optional[int] = None, fixed: Optional[int] = None, make_linear: Optional[int] = None) -> 'VideoNode': ...
+    def ClipToProp(self, mclip: 'VideoNode', prop: Optional[DataType] = None) -> 'VideoNode': ...
+    def Convolution(self, matrix: SingleAndSequence[float], bias: Optional[float] = None, divisor: Optional[float] = None, planes: Optional[SingleAndSequence[int]] = None, saturate: Optional[int] = None, mode: Optional[DataType] = None) -> 'VideoNode': ...
+    def CopyFrameProps(self, prop_src: 'VideoNode') -> 'VideoNode': ...
+    def Crop(self, left: Optional[int] = None, right: Optional[int] = None, top: Optional[int] = None, bottom: Optional[int] = None) -> 'VideoNode': ...
+    def CropAbs(self, width: int, height: int, left: Optional[int] = None, top: Optional[int] = None, x: Optional[int] = None, y: Optional[int] = None) -> 'VideoNode': ...
+    def CropRel(self, left: Optional[int] = None, right: Optional[int] = None, top: Optional[int] = None, bottom: Optional[int] = None) -> 'VideoNode': ...
+    def Deflate(self, planes: Optional[SingleAndSequence[int]] = None, threshold: Optional[float] = None) -> 'VideoNode': ...
+    def DeleteFrames(self, frames: SingleAndSequence[int]) -> 'VideoNode': ...
+    def DoubleWeave(self, tff: Optional[int] = None) -> 'VideoNode': ...
+    def DuplicateFrames(self, frames: SingleAndSequence[int]) -> 'VideoNode': ...
+    def Expr(self, expr: SingleAndSequence[DataType], format: Optional[int] = None) -> 'VideoNode': ...
+    def FlipHorizontal(self) -> 'VideoNode': ...
+    def FlipVertical(self) -> 'VideoNode': ...
+    def FrameEval(self, eval: VSMapValueCallback[VideoNode], prop_src: Optional[SingleAndSequence[VideoNode]] = None, clip_src: Optional[SingleAndSequence[VideoNode]] = None) -> 'VideoNode': ...
+    def FreezeFrames(self, first: SingleAndSequence[int], last: SingleAndSequence[int], replacement: SingleAndSequence[int]) -> 'VideoNode': ...
+    def Inflate(self, planes: Optional[SingleAndSequence[int]] = None, threshold: Optional[float] = None) -> 'VideoNode': ...
+    def Interleave(self, extend: Optional[int] = None, mismatch: Optional[int] = None, modify_duration: Optional[int] = None) -> 'VideoNode': ...
+    def Invert(self, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def InvertMask(self, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Levels(self, min_in: Optional[SingleAndSequence[float]] = None, max_in: Optional[SingleAndSequence[float]] = None, gamma: Optional[SingleAndSequence[float]] = None, min_out: Optional[SingleAndSequence[float]] = None, max_out: Optional[SingleAndSequence[float]] = None, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Limiter(self, min: Optional[SingleAndSequence[float]] = None, max: Optional[SingleAndSequence[float]] = None, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Loop(self, times: Optional[int] = None) -> 'VideoNode': ...
+    def Lut(self, planes: Optional[SingleAndSequence[int]] = None, lut: Optional[SingleAndSequence[int]] = None, lutf: Optional[SingleAndSequence[float]] = None, function: Optional[VSMapValueCallback[Union[int, float]]] = None, bits: Optional[int] = None, floatout: Optional[int] = None) -> 'VideoNode': ...
+    def Lut2(self, clipb: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, lut: Optional[SingleAndSequence[int]] = None, lutf: Optional[SingleAndSequence[float]] = None, function: Optional[VSMapValueCallback[Union[int, float]]] = None, bits: Optional[int] = None, floatout: Optional[int] = None) -> 'VideoNode': ...
+    def MakeDiff(self, clipb: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def MaskedMerge(self, clipb: 'VideoNode', mask: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None, first_plane: Optional[int] = None, premultiplied: Optional[int] = None) -> 'VideoNode': ...
+    def Maximum(self, planes: Optional[SingleAndSequence[int]] = None, threshold: Optional[float] = None, coordinates: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Median(self, planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Merge(self, clipb: 'VideoNode', weight: Optional[SingleAndSequence[float]] = None) -> 'VideoNode': ...
+    def MergeDiff(self, clipb: 'VideoNode', planes: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def Minimum(self, planes: Optional[SingleAndSequence[int]] = None, threshold: Optional[float] = None, coordinates: Optional[SingleAndSequence[int]] = None) -> 'VideoNode': ...
+    def ModifyFrame(self, clips: SingleAndSequence['VideoNode'], selector: VSMapValueCallback[VideoFrame]) -> 'VideoNode': ...
+    def PEMVerifier(self, upper: Optional[SingleAndSequence[float]] = None, lower: Optional[SingleAndSequence[float]] = None) -> 'VideoNode': ...
+    def PlaneStats(self, clipb: Optional['VideoNode'] = None, plane: Optional[int] = None, prop: Optional[DataType] = None) -> 'VideoNode': ...
+    def PreMultiply(self, alpha: 'VideoNode') -> 'VideoNode': ...
+    def Prewitt(self, planes: Optional[SingleAndSequence[int]] = None, scale: Optional[float] = None) -> 'VideoNode': ...
+    def PropToClip(self, prop: Optional[DataType] = None) -> 'VideoNode': ...
+    def RemoveFrameProps(self, props: Optional[SingleAndSequence[DataType]] = None) -> 'VideoNode': ...
+    def Reverse(self) -> 'VideoNode': ...
+    def SelectEvery(self, cycle: int, offsets: SingleAndSequence[int], modify_duration: Optional[int] = None) -> 'VideoNode': ...
+    def SeparateFields(self, tff: Optional[int] = None, modify_duration: Optional[int] = None) -> 'VideoNode': ...
+    def SetFieldBased(self, value: int) -> 'VideoNode': ...
+    def SetFrameProp(self, prop: DataType, intval: Optional[SingleAndSequence[int]] = None, floatval: Optional[SingleAndSequence[float]] = None, data: Optional[SingleAndSequence[DataType]] = None) -> 'VideoNode': ...
+    def SetFrameProps(self, **kwargs: Any) -> 'VideoNode': ...
+    def SetVideoCache(self, mode: Optional[int] = None, fixedsize: Optional[int] = None, maxsize: Optional[int] = None, maxhistory: Optional[int] = None) -> None: ...
+    def ShufflePlanes(self, planes: SingleAndSequence[int], colorfamily: int) -> 'VideoNode': ...
+    def Sobel(self, planes: Optional[SingleAndSequence[int]] = None, scale: Optional[float] = None) -> 'VideoNode': ...
+    def Splice(self, mismatch: Optional[int] = None) -> 'VideoNode': ...
+    def SplitPlanes(self) -> SingleAndSequence['VideoNode']: ...
+    def StackHorizontal(self) -> 'VideoNode': ...
+    def StackVertical(self) -> 'VideoNode': ...
+    def Transpose(self) -> 'VideoNode': ...
+    def Trim(self, first: Optional[int] = None, last: Optional[int] = None, length: Optional[int] = None) -> 'VideoNode': ...
+    def Turn180(self) -> 'VideoNode': ...
+
+class _Plugin_std_AudioNode_Bound(Plugin):
+    """This class implements the module definitions for the "std" VapourSynth plugin.\n\n*This class cannot be imported.*"""
+    def AssumeSampleRate(self, src: Optional['AudioNode'] = None, samplerate: Optional[int] = None) -> 'AudioNode': ...
+    def AudioGain(self, gain: Optional[SingleAndSequence[float]] = None) -> 'AudioNode': ...
+    def AudioLoop(self, times: Optional[int] = None) -> 'AudioNode': ...
+    def AudioMix(self, matrix: SingleAndSequence[float], channels_out: SingleAndSequence[int]) -> 'AudioNode': ...
+    def AudioReverse(self) -> 'AudioNode': ...
+    def AudioSplice(self) -> 'AudioNode': ...
+    def AudioTrim(self, first: Optional[int] = None, last: Optional[int] = None, length: Optional[int] = None) -> 'AudioNode': ...
+    def BlankAudio(self, channels: Optional[int] = None, bits: Optional[int] = None, sampletype: Optional[int] = None, samplerate: Optional[int] = None, length: Optional[int] = None, keep: Optional[int] = None) -> 'AudioNode': ...
+    def SetAudioCache(self, mode: Optional[int] = None, fixedsize: Optional[int] = None, maxsize: Optional[int] = None, maxhistory: Optional[int] = None) -> None: ...
+    def ShuffleChannels(self, channels_in: SingleAndSequence[int], channels_out: SingleAndSequence[int]) -> 'AudioNode': ...
+    def SplitChannels(self) -> SingleAndSequence['AudioNode']: ...
+
+# end implementation
+
+
+
+class RawNode:
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    def get_frame(self, n: int) -> RawFrame: ...
+
+    @overload
+    def get_frame_async(self, n: int, cb: None = None) -> _Future[RawFrame]: ...
+
+    @overload
+    def get_frame_async(self, n: int, cb: Callable[[Union[RawFrame, None], Union[Exception, None]], None]) -> None: ...
+
+    def frames(
+        self, prefetch: Union[int, None] = None, backlog: Union[int, None] = None, close: bool = False
+    ) -> Iterator[RawFrame]: ...
+
+    def set_output(self, index: int = 0) -> None: ...
+
+    def is_inspectable(self, version: Union[int, None] = None) -> bool: ...
+
+    if not TYPE_CHECKING:
+        @property
+        def _node_name(self) -> str: ...
+
+        @property
+        def _name(self) -> str: ...
+
+        @property
+        def _inputs(self) -> Dict[str, _VapourSynthMapValue]: ...
+
+        @property
+        def _timings(self) -> int: ...
+
+        @property
+        def _mode(self) -> FilterMode: ...
+
+        @property
+        def _dependencies(self): ...
+
+    @overload
+    def __eq__(self: 'SelfRawNode', other: 'SelfRawNode', /) -> bool: ...  # type: ignore[misc]
+
+    @overload
+    def __eq__(self: 'SelfRawNode', other: Any, /) -> Literal[False]: ...
+
+    def __add__(self: 'SelfRawNode', other: 'SelfRawNode', /) -> 'SelfRawNode': ...
+
+    def __radd__(self: 'SelfRawNode', other: 'SelfRawNode', /) -> 'SelfRawNode': ...
+
+    def __mul__(self: 'SelfRawNode', other: int) -> 'SelfRawNode': ...
+
+    def __rmul__(self: 'SelfRawNode', other: int) -> 'SelfRawNode': ...
+
+    def __getitem__(self: 'SelfRawNode', index: Union[int, slice], /) -> 'SelfRawNode': ...
+
+    def __getattr__(self, name: str) -> Plugin: ...
+
+    def __len__(self) -> int: ...
+
+
+SelfRawNode = TypeVar('SelfRawNode', bound=RawNode)
+
+
+class VideoNode(RawNode):
+    format: Union[VideoFormat, None]
+
+    width: int
+    height: int
+
+    fps_num: int
+    fps_den: int
+
+    fps: Fraction
+
+    num_frames: int
+
+    def set_output(
+        self, index: int = 0, alpha: Union['VideoNode', None] = None, alt_output: Literal[0, 1, 2] = 0
+    ) -> None: ...
+
+    def output(
+        self, fileobj: BinaryIO, y4m: bool = False, progress_update: object = None, prefetch: int = 0, backlog: int = -1
+    ) -> None: ...
+
+    def get_frame(self, n: int) -> VideoFrame: ...
+
+    @overload  # type: ignore[override]
+    def get_frame_async(self, n: int, cb: None = None) -> _Future[VideoFrame]: ...
+
+    @overload
+    def get_frame_async(self, n: int, cb: Callable[[Union[VideoFrame, None], Union[Exception, None]], None]) -> None: ...
+
+    def frames(
+        self, prefetch: Union[int, None] = None, backlog: Union[int, None] = None, close: bool = False
+    ) -> Iterator[VideoFrame]: ...
+
+    # instance_bound_VideoNode: descale
+    @property
+    def descale(self) -> _Plugin_descale_VideoNode_Bound:
+        """Undo linear interpolation"""
+    # end instance
+    # instance_bound_VideoNode: fmtc
+    @property
+    def fmtc(self) -> _Plugin_fmtc_VideoNode_Bound:
+        """Format converter, r28"""
+    # end instance
+    # instance_bound_VideoNode: placebo
+    @property
+    def placebo(self) -> _Plugin_placebo_VideoNode_Bound:
+        """libplacebo plugin for VapourSynth"""
+    # end instance
+    # instance_bound_VideoNode: resize
+    @property
+    def resize(self) -> _Plugin_resize_VideoNode_Bound:
+        """VapourSynth Resize"""
+    # end instance
+    # instance_bound_VideoNode: std
+    @property
+    def std(self) -> _Plugin_std_VideoNode_Bound:
+        """VapourSynth Core Functions"""
+    # end instance
+
+
+class AudioNode(RawNode):
+    sample_type: object
+    bits_per_sample: int
+    bytes_per_sample: int
+
+    channel_layout: int
+    num_channels: int
+
+    sample_rate: int
+    num_samples: int
+
+    num_frames: int
+
+    def get_frame(self, n: int) -> AudioFrame: ...
+
+    @overload  # type: ignore[override]
+    def get_frame_async(self, n: int, cb: None = None) -> _Future[AudioFrame]: ...
+
+    @overload
+    def get_frame_async(self, n: int, cb: Callable[[Union[AudioFrame, None], Union[Exception, None]], None]) -> None: ...
+
+    def frames(
+        self, prefetch: Union[int, None] = None, backlog: Union[int, None] = None, close: bool = False
+    ) -> Iterator[AudioFrame]: ...
+
+    # instance_bound_AudioNode: std
+    @property
+    def std(self) -> _Plugin_std_AudioNode_Bound:
+        """VapourSynth Core Functions"""
+    # end instance
+
+
+class VSMapValueCallbackData:
+    def __init__(
+        self, node: RawNode, env: EnvironmentData,
+        callback: Union[Callable[[Union[RawFrame, None], Union[Exception, None]], None], None] = None
+    ) -> None: ...
+
+    def receive(self, n: int, result: Union[RawFrame, Exception]) -> None: ...
+
+
+class LogHandle:
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+
+class PythonVSScriptLoggingBridge(Handler):
+
+    def __init__(self, parent: Handler, level: int = LogLevelUnset) -> None: ...
+
+    def emit(self, record: LogRecord) -> None: ...
+
+
+class Function:
+    plugin: 'Plugin'
+    name: str
+    signature: str
+    return_signature: str
+
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    def __call__(self, *args: _VapourSynthMapValue, **kwargs: _VapourSynthMapValue) -> _VapourSynthMapValue: ...
+
+    @property
+    def __signature__(self) -> Signature: ...
+
+
+class Plugin:
+    identifier: str
+    namespace: str
+    name: str
+
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    def __getattr__(self, name: str) -> Function: ...
+
+    def functions(self) -> Iterator[Function]: ...
+
+
+class Core:
+    def __init__(self) -> NoReturn: ...  # type: ignore[misc]
+
+    @property
+    def num_threads(self) -> int: ...
+
+    @num_threads.setter
+    def num_threads(self) -> None: ...
+
+    @property
+    def max_cache_size(self) -> int: ...
+
+    @max_cache_size.setter
+    def max_cache_size(self) -> None: ...
+
+    @property
+    def flags(self) -> int: ...
+
+    def plugins(self) -> Iterator[Plugin]: ...
+
+    def query_video_format(
+        self, color_family: ColorFamily, sample_type: SampleType, bits_per_sample: int, subsampling_w: int = 0,
+        subsampling_h: int = 0
+    ) -> VideoFormat: ...
+
+    def get_video_format(self, id: Union[VideoFormat, int, PresetFormat]) -> VideoFormat: ...
+
+    def create_video_frame(self, format: VideoFormat, width: int, height: int) -> VideoFrame: ...
+
+    def log_message(self, message_type: MessageType, message: str) -> None: ...
+
+    def add_log_handler(self, handler_func: Callable[[MessageType, str], None]) -> LogHandle: ...
+
+    def remove_log_handler(self, handle: LogHandle) -> None: ...
+
+    def version(self) -> str: ...
+
+    def version_number(self) -> int: ...
+
+    def __getattr__(self, name: str) -> Plugin: ...
+
+    # instance_bound_Core: descale
+    @property
+    def descale(self) -> _Plugin_descale_Core_Bound:
+        """Undo linear interpolation"""
+    # end instance
+    # instance_bound_Core: fmtc
+    @property
+    def fmtc(self) -> _Plugin_fmtc_Core_Bound:
+        """Format converter, r28"""
+    # end instance
+    # instance_bound_Core: placebo
+    @property
+    def placebo(self) -> _Plugin_placebo_Core_Bound:
+        """libplacebo plugin for VapourSynth"""
+    # end instance
+    # instance_bound_Core: resize
+    @property
+    def resize(self) -> _Plugin_resize_Core_Bound:
+        """VapourSynth Resize"""
+    # end instance
+    # instance_bound_Core: std
+    @property
+    def std(self) -> _Plugin_std_Core_Bound:
+        """VapourSynth Core Functions"""
+    # end instance
+
+
+class _CoreProxy(Core):
+    @property
+    def core(self) -> Core: ...
+
+
+core: _CoreProxy
