@@ -1,53 +1,19 @@
 from __future__ import annotations
 
+from typing import Any
+
+from vstools import CustomValueError, FuncExceptT
+
 __all__ = [
-    'ReservedMatrixError',
-    'UndefinedMatrixError',
     'UnknownKernelError',
-    'UnsupportedMatrixError',
-    'VideoPropError',
 ]
 
 
-class UndefinedMatrixError(ValueError):
-    """Raised when an undefined matrix is passed."""
-
-
-class ReservedMatrixError(PermissionError):
-    """Raised when a reserved matrix is requested."""
-
-
-class UnsupportedMatrixError(ValueError):
-    """Raised when an unsupported matrix is passed."""
-
-
-class UndefinedTransferError(ValueError):
-    """Raised when an undefined transfer is passed."""
-
-
-class ReservedTransferError(PermissionError):
-    """Raised when a reserved transfer is requested."""
-
-
-class UnsupportedTransferError(ValueError):
-    """Raised when an unsupported transfer is passed."""
-
-
-class UndefinedPrimariesError(ValueError):
-    """Raised when an undefined primaries value is passed."""
-
-
-class ReservedPrimariesError(PermissionError):
-    """Raised when reserved primaries are requested."""
-
-
-class UnsupportedPrimariesError(ValueError):
-    """Raised when a unsupported primaries value is passed."""
-
-
-class UnknownKernelError(ValueError):
+class UnknownKernelError(CustomValueError):
     """Raised when an unknown kernel is passed."""
 
-
-class VideoPropError(KeyError):
-    """Raised when there was an issue with a VideoProp."""
+    def __init__(
+        self, function: FuncExceptT, kernel: str, message: str = 'Unknown kernel: {kernel}!',
+        **kwargs: Any
+    ) -> None:
+        super().__init__(message, function, kernel=kernel, **kwargs)
